requirements.txt
requests==2.31.0
jsonschema==4.20.0
streamlit==1.31.0
streamlit-drawable-canvas==0.9.3
Pillow==10.1.0 
openai